+++ src/StardewModdingAPI/Events/EventArgsCurrentLocationChanged.cs
@@ -9,10 +9,10 @@ namespace StardewModdingAPI.Events
         /*********
         ** Accessors
         *********/
-        /// <summary>The player's previous location.</summary>
+        /// <summary>The player's current location.</summary>
         public GameLocation NewLocation { get; private set; }
 
-        /// <summary>The player's current location.</summary>
+        /// <summary>The player's previous location.</summary>
         public GameLocation PriorLocation { get; private set; }